tests/wpt/tests/navigation-api/navigation-methods/reload-no-args.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    const navState = { key: "value" };
    i.contentWindow.navigation.navigate("#1", { state: navState }).committed.then(t.step_func(() => {
      // Make sure that state setting worked
      assert_equals(i.contentWindow.navigation.currentEntry.getState().key, "value");
      assert_not_equals(i.contentWindow.navigation.currentEntry.getState(), navState);

      let start_url = i.contentWindow.location.href;
      let start_key = i.contentWindow.navigation.currentEntry.key;
      let start_id = i.contentWindow.navigation.currentEntry.id;
      let start_state = i.contentWindow.navigation.currentEntry.getState();
      let onnavigate_called = false;
      let promise_settled = false;
      i.contentWindow.navigation.onnavigate = t.step_func(e => {
        onnavigate_called = true;
        assert_equals(e.info, undefined);
        assert_equals(e.navigationType, "reload");
        assert_equals(e.destination.getState().key, "value");
        assert_not_equals(e.destination.getState(), start_state);
      });
      i.contentWindow.navigation.reload().committed.finally(() => {
        promise_settled = true;
      });
      i.onload = t.step_func(() => {
        assert_true(onnavigate_called);
        assert_equals(i.contentWindow.location.href, start_url);
        assert_equals(i.contentWindow.navigation.currentEntry.key, start_key);
        assert_equals(i.contentWindow.navigation.currentEntry.id, start_id);
        assert_equals(i.contentWindow.navigation.currentEntry.getState().key, "value");
        assert_not_equals(i.contentWindow.navigation.currentEntry.getState(), start_state);
        assert_equals(i.contentWindow.navigation.entries().length, 2);

        t.step_timeout(t.step_func_done(() => {
          assert_equals(promise_settled, false);
        }), 0);
      });
    }));
  });
}, "reload() variant with no state or info");
</script>
